# src/doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/wcag/url": {
      "get": {
        "summary": "Check WCAG directives to a website URL and return a report",
        "description": "Based on the WCAG 2 Overview (https://www.w3.org/WAI/standards-guidelines/wcag/)",
        "parameters": [
          {
            "name": "website_url",
            "in": "query",
            "description": "Indicates the website URL to test",
            "required": true,
            "type": "string"
          },
          {
            "name": "webdriver_url",
            "in": "query",
            "description": "Webdriver URL",
            "required": true,
            "type": "string"
          },
          {
            "name": "wcag_version",
            "in": "query",
            "description": "Indicates the WCAG version to test",
            "required": true,
            "type": "string",
            "enum": [
              "*",
              "2.0",
              "2.1"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "WCAG distinguishable contrasts results",
            "schema": {
              "$ref": "#/definitions/WCAGTextContrastAnswerList"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/wcag/text": {
      "get": {
        "summary": "Check text contrast based on the WCAG norm",
        "description": "Based on the Contrast subsection the Distinguishable section of the WCAG 2 Overview (https://www.w3.org/WAI/standards-guidelines/wcag/)",
        "parameters": [
          {
            "name": "wcag_version",
            "in": "query",
            "description": "Indicates the WCAG version to test",
            "required": true,
            "type": "string",
            "enum": [
              "*",
              "2.0",
              "2.1"
            ]
          },
          {
            "name": "background_color",
            "in": "query",
            "description": "Background color (HEX, RGB, RGBA, HSL & HSLA)",
            "required": true,
            "type": "string",
            "pattern": "/^(#?([(a-f|A-F)d]{3,4}|[(a-f|A-F)d]{6}|[(a-f|A-F)d]{8})|rgb((0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|255|25[0-4]|2[0-4]d|1dd|0?d?d))|rgba((0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|0?.d|1(.0)?))|hsl((0|360|35d|3[0-4]d|[12]dd|0?d?d),(0|100|d{1,2})%,(0|100|d{1,2})%)|hsla((0|360|35d|3[0-4]d|[12]dd|0?d?d),(0|100|d{1,2})%,(0|100|d{1,2})%,(0?.d|1(.0)?)))$/"
          },
          {
            "name": "text_color",
            "in": "query",
            "description": "Text color (HEX, RGB, RGBA, HSL & HSLA)",
            "required": true,
            "type": "string",
            "pattern": "/^(#?([(a-f|A-F)d]{3,4}|[(a-f|A-F)d]{6}|[(a-f|A-F)d]{8})|rgb((0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|255|25[0-4]|2[0-4]d|1dd|0?d?d))|rgba((0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|255|25[0-4]|2[0-4]d|1dd|0?d?d),(0|0?.d|1(.0)?))|hsl((0|360|35d|3[0-4]d|[12]dd|0?d?d),(0|100|d{1,2})%,(0|100|d{1,2})%)|hsla((0|360|35d|3[0-4]d|[12]dd|0?d?d),(0|100|d{1,2})%,(0|100|d{1,2})%,(0?.d|1(.0)?)))$/"
          },
          {
            "name": "text_size",
            "in": "query",
            "description": "Indicates the text size in points",
            "required": true,
            "type": "number"
          },
          {
            "name": "text_bold",
            "in": "query",
            "description": "Indicates if the text is in bold",
            "required": true,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "WCAG distinguishable contrasts results",
            "schema": {
              "$ref": "#/definitions/WCAGTextContrastAnswer"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "WCAGContrastAnswer": {
      "type": "object",
      "properties": {
        "wcagVersion": {
          "type": "string",
          "example": "2.0"
        },
        "value": {
          "type": "number",
          "example": 21
        },
        "minimum": {
          "type": "boolean",
          "example": true
        },
        "enhanced": {
          "type": "boolean",
          "example": true
        },
        "isLargeText": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "WCAGTextContrastAnswer": {
      "type": "object",
      "properties": {
        "contrasts": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "wcagVersion": {
                "type": "string",
                "example": "2.0"
              },
              "value": {
                "type": "number",
                "example": 21
              },
              "minimum": {
                "type": "boolean",
                "example": true
              },
              "enhanced": {
                "type": "boolean",
                "example": true
              },
              "isLargeText": {
                "type": "boolean",
                "example": false
              }
            }
          }
        },
        "textBackgroundColor": {
          "type": "string",
          "example": "#FFF"
        },
        "textColor": {
          "type": "string",
          "example": "#000"
        },
        "textSize": {
          "type": "number",
          "example": 12
        },
        "textBold": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "WCAGTextContrastAnswerList": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "contrasts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "wcagVersion": {
                  "type": "string",
                  "example": "2.0"
                },
                "value": {
                  "type": "number",
                  "example": 21
                },
                "minimum": {
                  "type": "boolean",
                  "example": true
                },
                "enhanced": {
                  "type": "boolean",
                  "example": true
                },
                "isLargeText": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          },
          "textBackgroundColor": {
            "type": "string",
            "example": "#FFF"
          },
          "textColor": {
            "type": "string",
            "example": "#000"
          },
          "textSize": {
            "type": "number",
            "example": 12
          },
          "textBold": {
            "type": "boolean",
            "example": false
          }
        }
      }
    }
  }
}